15_PDF_RAG/requirements.txt
pdfplumber
faiss-cpu
requests
numpy
streamlit
